refractor: quizz controller and quizz play controller
